requirements.txt
amqp==5.2.0
asgiref==3.6.0
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
defusedxml==0.7.1
Django==3.2.13
django-allauth==0.56.1
django-ckeditor==6.7.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-js-asset==2.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
isort==5.12.0
kombu==5.3.4
oauthlib==3.2.2
packaging==23.2
Pillow==9.5.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.7
pycparser==2.21
pydantic==1.10.8
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stripe==6.2.0
typing_extensions==4.6.2
tzdata==2023.4
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
